+++ cyEngine/cyUtilities/include/cyZiggurat.h
@@ -23,7 +23,7 @@ namespace CYLLENE_SDK
 
 
     uint32 cong_seeded(uint32& jcong) {
-      uint32_t value;
+      uint32 value;
 
       jcong = 69069 * (jcong)+1234567;
 
@@ -33,7 +33,7 @@ namespace CYLLENE_SDK
     }
     
     uint32 kiss_seeded(uint32& jcong, uint32& jsr, uint32& w, uint32& z) {
-      uint32_t value;
+      uint32 value;
 
       value = (mwc_seeded(w, z) ^ cong_seeded(jcong)) + shr3_seeded(jsr);
 
@@ -41,7 +41,7 @@ namespace CYLLENE_SDK
     }
 
     uint32 mwc_seeded(uint32& w, uint32& z) {
-      uint32_t value;
+      uint32 value;
 
       z = 36969 * (z & 65535) + (z >> 16);
       w = 18000 * (w & 65535) + (w >> 16);
@@ -52,8 +52,8 @@ namespace CYLLENE_SDK
     }
 
     float r4_exp(uint32& jsr, uint32 ke[256], float fe[256], float we[256]) {
-      uint32_t iz;
-      uint32_t jz;
+      uint32 iz;
+      uint32 jz;
       float value;
       float x;
 
@@ -61,16 +61,16 @@ namespace CYLLENE_SDK
       iz = (jz & 255);
 
       if (jz < ke[iz]) {
-        value = (float)(jz)*we[iz];
+        value = static_cast<float>(jz)*we[iz];
       }
       else {
         for (; ; ) {
           if (iz == 0) {
-            value = 7.69711 - log(r4_uni(jsr));
+            value = static_cast<float>(7.69711 - log(r4_uni(jsr)));
             break;
           }
 
-          x = (float)(jz)*we[iz];
+          x = static_cast<float>(jz)*we[iz];
 
           if (fe[iz] + r4_uni(jsr) * (fe[iz - 1] - fe[iz]) < exp(-x)) {
             value = x;
@@ -81,7 +81,7 @@ namespace CYLLENE_SDK
           iz = (jz & 255);
 
           if (jz < ke[iz]) {
-            value = (float)(jz)*we[iz];
+            value = static_cast<float>(jz)*we[iz];
             break;
           }
         }
@@ -99,44 +99,44 @@ namespace CYLLENE_SDK
 
       q = ve / exp(-de);
 
-      ke[0] = (uint32_t)((de / q) * m2);
+      ke[0] = static_cast<uint32>((de / q) * m2);
       ke[1] = 0;
 
-      we[0] = (float)(q / m2);
-      we[255] = (float)(de / m2);
+      we[0] = static_cast<float>(q / m2);
+      we[255] = static_cast<float>(de / m2);
 
       fe[0] = 1.0;
-      fe[255] = (float)(exp(-de));
+      fe[255] = static_cast<float>(exp(-de));
 
       for (i = 254; 1 <= i; i--) {
         de = -log(ve / de + exp(-de));
-        ke[i + 1] = (uint32_t)((de / te) * m2);
+        ke[i + 1] = static_cast<uint32>((de / te) * m2);
         te = de;
-        fe[i] = (float)(exp(-de));
-        we[i] = (float)(de / m2);
+        fe[i] = static_cast<float>(exp(-de));
+        we[i] = static_cast<float>(de / m2);
       }
       return;
     }
     
     float r4_nor(uint32& jsr, uint32 kn[128], float fn[128], float wn[128]) {
-      int hz;
-      uint32_t iz;
-      const float r = 3.442620;
+      int32 hz;
+      uint32 iz;
+      const float r = 3.442620f;
       float value;
       float x;
       float y;
 
-      hz = (int)shr3_seeded(jsr);
+      hz = static_cast<int32>(shr3_seeded(jsr));
       iz = (hz & 127);
 
       if (fabs(hz) < kn[iz]) {
-        value = (float)(hz)*wn[iz];
+        value = static_cast<float>(hz)*wn[iz];
       }
       else {
         for (; ; ) {
           if (iz == 0) {
             for (; ; ) {
-              x = -0.2904764 * log(r4_uni(jsr));
+              x = static_cast<float>(- 0.2904764 * log(r4_uni(jsr)));
               y = -log(r4_uni(jsr));
               if (x * x <= y + y) {
                 break;
@@ -152,7 +152,7 @@ namespace CYLLENE_SDK
             break;
           }
 
-          x = (float)(hz)*wn[iz];
+          x = static_cast<float>(hz)*wn[iz];
 
           if (fn[iz] + r4_uni(jsr) * (fn[iz - 1] - fn[iz]) < 
               exp(-0.5 * x * x)) {
@@ -160,11 +160,11 @@ namespace CYLLENE_SDK
             break;
           }
 
-          hz = (int)shr3_seeded(jsr);
+          hz = static_cast<int32>(shr3_seeded(jsr));
           iz = (hz & 127);
 
           if (fabs(hz) < kn[iz]) {
-            value = (float)(hz)*wn[iz];
+            value = static_cast<float>(hz)*wn[iz];
             break;
           }
         }
@@ -183,28 +183,28 @@ namespace CYLLENE_SDK
 
       q = vn / exp(-0.5 * dn * dn);
 
-      kn[0] = (uint32_t)((dn / q) * m1);
+      kn[0] = static_cast<uint32>((dn / q) * m1);
       kn[1] = 0;
 
-      wn[0] = (float)(q / m1);
-      wn[127] = (float)(dn / m1);
+      wn[0] = static_cast<float>(q / m1);
+      wn[127] = static_cast<float>(dn / m1);
 
       fn[0] = 1.0;
-      fn[127] = (float)(exp(-0.5 * dn * dn));
+      fn[127] = static_cast<float>(exp(-0.5 * dn * dn));
 
       for (i = 126; 1 <= i; i--) {
         dn = sqrt(-2.0 * log(vn / dn + exp(-0.5 * dn * dn)));
-        kn[i + 1] = (uint32_t)((dn / tn) * m1);
+        kn[i + 1] = static_cast<uint32>((dn / tn) * m1);
         tn = dn;
-        fn[i] = (float)(exp(-0.5 * dn * dn));
-        wn[i] = (float)(dn / m1);
+        fn[i] = static_cast<float>(exp(-0.5 * dn * dn));
+        wn[i] = static_cast<float>(dn / m1);
       }
 
       return;
     }
     
     float r4_uni(uint32& jsr) {
-      uint32_t jsr_input;
+      uint32 jsr_input;
       float value;
 
       jsr_input = jsr;
@@ -213,14 +213,15 @@ namespace CYLLENE_SDK
       jsr = (jsr ^ (jsr >> 17));
       jsr = (jsr ^ (jsr << 5));
 
-      value = fmod(0.5 + (float)(jsr_input + jsr) / 65536.0 / 65536.0, 1.0);
+      value = static_cast<float>(fmod(0.5 + static_cast<float>(jsr_input + jsr) / 65536.0 / 65536.0, 
+                                      1.0));
 
       return value;
     }
     
     uint32 shr3_seeded(uint32& jsr) {
-      uint32_t jsr_input;
-      uint32_t value;
+      uint32 jsr_input;
+      uint32 value;
 
       jsr_input = jsr;
 
@@ -234,7 +235,7 @@ namespace CYLLENE_SDK
     }
     std::mt19937 m_generator;
     float m_fn[128];
-    uint32_t m_kn[128];
+    uint32 m_kn[128];
     float m_wn[128];
   };
